Add container/presentation structure
