+++ app/src/components/ContactForm/index.js
@@ -1,101 +1,32 @@
 import PropTypes from 'prop-types';
 import {
-  useState, useEffect, forwardRef, useImperativeHandle,
+  forwardRef,
 } from 'react';
 import FormGroup from '../FormGroup';
 import { Form, ButtonContainer } from './styles';
 
-import isEmailValid from '../../utils/isEmailValid';
-import formatPhone from '../../utils/formatPhone';
-import categoriesService from '../../services/CategoriesService';
-
 import Input from '../Input';
 import Select from '../Select';
 import Button from '../Button';
-import useErros from '../../hooks/useErrors';
-import useSafeAsyncState from '../../hooks/useSafeAsyncState';
+import useContactForm from './useContactForm';
 
 const ContactForm = forwardRef(({ buttonLabel, onSubmit }, ref) => {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [phone, setPhone] = useState('');
-  const [categoryId, setCategoryId] = useState('');
-  const [categories, setCategories] = useSafeAsyncState([]);
-  const [isLoadingCategories, setIsLoadingCategories] = useSafeAsyncState(true);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
   const {
-    setError, removeError, getErrorMessageByFieldName, errors,
-  } = useErros();
-
-  const isFormValid = (name && !errors.length);
-
-  useImperativeHandle(ref, () => ({
-    setFieldsValues(contact) {
-      setName(contact.name ?? '');
-      setEmail(contact.email ?? '');
-      setPhone(formatPhone(contact.phone ?? ''));
-      setCategoryId(contact.category.id ?? '');
-    },
-    resetFields() {
-      setName('');
-      setEmail('');
-      setPhone('');
-      setCategoryId('');
-    },
-  }), []);
-
-  useEffect(() => {
-    async function loadCategories() {
-      try {
-        const categoriesList = await categoriesService.listCategories();
-        setCategories(categoriesList);
-      } catch { } finally {
-        setIsLoadingCategories(false);
-      }
-    }
-
-    loadCategories();
-  }, [setCategories, setIsLoadingCategories]);
-
-  function handleNameChange(e) {
-    setName(e.target.value);
-
-    if (!e.target.value) {
-      setError({ field: 'name', message: 'Nome é obrigatório.' });
-    } else {
-      removeError('name');
-    }
-  }
-
-  function handleEmailChange(e) {
-    setEmail(e.target.value);
-
-    if (e.target.value && !isEmailValid(e.target.value)) {
-      setError({ field: 'email', message: 'E-mail é inválido.' });
-    } else {
-      removeError('email');
-    }
-  }
-
-  function handlePhoneChange(e) {
-    setPhone(formatPhone(e.target.value));
-  }
-
-  async function handleSubmit(e) {
-    e.preventDefault();
-
-    setIsSubmitting(true);
-
-    await onSubmit({
-      name,
-      email,
-      phone,
-      categoryId,
-    });
-
-    setIsSubmitting(false);
-  }
+    handleSubmit,
+    getErrorMessageByFieldName,
+    name,
+    handleNameChange,
+    isSubmitting,
+    email,
+    handleEmailChange,
+    phone,
+    handlePhoneChange,
+    isLoadingCategories,
+    categoryId,
+    setCategoryId,
+    categories,
+    isFormValid,
+  } = useContactForm(onSubmit, ref);
 
   return (
     <Form onSubmit={handleSubmit} noValidate>
