+++ app/src/components/Toast/ToastContainer/index.js
@@ -1,31 +1,9 @@
-import { useCallback, useEffect, useState } from 'react';
 import ToastMessage from '../ToastMessage';
 import { Container } from './styles';
-
-import { toastEventManager } from '../../../utils/toast';
+import useToastContainer from './useToastContainer';
 
 export default function ToastContainer() {
-  const [messages, setMessages] = useState([]);
-
-  useEffect(() => {
-    function handleAddToast({ type, text, duration }) {
-      setMessages((prev) => [
-        ...prev, {
-          id: Math.random(), type, text, duration,
-        },
-      ]);
-    }
-
-    toastEventManager.on('addtoast', handleAddToast);
-
-    return () => {
-      toastEventManager.removeListener('addtoast', handleAddToast);
-    };
-  }, []);
-
-  const handleRemoveMessage = useCallback((id) => {
-    setMessages((prev) => prev.filter((message) => message.id !== id));
-  }, []);
+  const { messages, handleRemoveMessage } = useToastContainer();
 
   return (
     <Container>
