+++ app/src/components/Toast/ToastMessage/index.js
@@ -1,25 +1,13 @@
 import PropTypes from 'prop-types';
 
-import { useEffect } from 'react';
 import { Container } from './styles';
 
 import xCircleIcon from '../../../assets/images/icons/x-circle.svg';
 import checkCircleIcon from '../../../assets/images/icons/check-circle.svg';
+import useToastMessage from './useToastMessage';
 
 export default function ToastMessage({ message, onRemoveMessage }) {
-  function handleRemoveToast() {
-    onRemoveMessage(message.id);
-  }
-
-  useEffect(() => {
-    const timeoutId = setTimeout(() => {
-      onRemoveMessage(message.id);
-    }, message.duration || 7000);
-
-    return () => {
-      clearTimeout(timeoutId);
-    };
-  }, [message, onRemoveMessage]);
+  const { handleRemoveToast } = useToastMessage(message, onRemoveMessage);
 
   return (
     <Container
