+++ app/src/components/Toast/ToastMessage/useToastMessage.js
@@ -1,0 +1,21 @@
+import { useEffect } from 'react';
+
+export default function useToastMessage(message, onRemoveMessage) {
+  function handleRemoveToast() {
+    onRemoveMessage(message.id);
+  }
+
+  useEffect(() => {
+    const timeoutId = setTimeout(() => {
+      onRemoveMessage(message.id);
+    }, message.duration || 7000);
+
+    return () => {
+      clearTimeout(timeoutId);
+    };
+  }, [message, onRemoveMessage]);
+
+  return {
+    handleRemoveToast,
+  };
+}
