+++ app/src/pages/EditContact/index.js
@@ -1,58 +1,12 @@
-import { useEffect, useRef, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
 import PageHeader from '../../components/PageHeader';
 import ContactForm from '../../components/ContactForm';
-import ContactsService from '../../services/ContactsService';
 import Loader from '../../components/Loader';
-import toast from '../../utils/toast';
-import useSafeAsyncAction from '../../hooks/useSafeAsyncAction';
+import useEditContact from './useEditContact';
 
 export default function EditContact() {
-  const [isLoading, setIsLoading] = useState(true);
-  const [contactName, setContactName] = useState('');
-
-  const contactFormRef = useRef(null);
-
-  const { id } = useParams();
-  const navigate = useNavigate();
-  const safeAsyncAction = useSafeAsyncAction();
-
-  useEffect(() => {
-    async function loadContact() {
-      try {
-        const contactData = await ContactsService.getContactById(id);
-
-        safeAsyncAction(() => {
-          contactFormRef.current.setFieldsValues(contactData);
-
-          setIsLoading(false);
-          setContactName(contactData.name);
-        });
-      } catch {
-        safeAsyncAction(() => {
-          navigate('/', { replace: true });
-          toast({
-            type: 'error',
-            text: 'Contato não encontrado.',
-          });
-        });
-      }
-    }
-
-    loadContact();
-  }, [id, navigate, safeAsyncAction]);
-
-  async function handleSubmit(contact) {
-    try {
-      const contactData = await ContactsService.updateContact(id, contact);
-
-      setContactName(contactData.name);
-
-      toast({ type: 'success', text: 'Contato editado com sucesso', duration: 3000 });
-    } catch {
-      toast({ type: 'error', text: 'Erro ao editar contato' });
-    }
-  }
+  const {
+    isLoading, contactName, contactFormRef, handleSubmit,
+  } = useEditContact();
 
   return (
     <>
