+++ app/src/pages/Home/index.js
@@ -1,15 +1,10 @@
 /* eslint-disable react/jsx-one-expression-per-line */
-import {
-  useEffect, useMemo, useState, useCallback, useTransition,
-} from 'react';
 import { Link } from 'react-router-dom';
 
 import Loader from '../../components/Loader';
 import Divisor from '../../components/Divisor';
 import Button from '../../components/Button';
 import Modal from '../../components/Modal';
-import toast from '../../utils/toast';
-import ContactsService from '../../services/ContactsService';
 
 import {
   Container,
@@ -25,94 +20,27 @@ import sad from '../../assets/images/icons/sad.svg';
 import emptyBox from '../../assets/images/icons/empty-box.svg';
 import magnifierQuestion from '../../assets/images/icons/magnifier-question.svg';
 import ContactsList from '../../components/ContactsList';
+import useHome from './useHome';
 
 export default function Home() {
-  const [contacts, setContacts] = useState([]);
-  const [orderBy, setOrderBy] = useState('asc');
-  const [isLoading, setIsLoading] = useState(true);
-  const [hasError, setHasError] = useState(false);
-  const [isDeleteModalVisible, setIsDeleteModalVisible] = useState(false);
-  const [contactBeingDeleted, setContactBeingDeleted] = useState(null);
-  const [isLoadingDelete, setIsLoadingDelete] = useState(false);
-
-  const [searchTerm, setSearchTerm] = useState('');
-  const [DeferredSearchTerm, setDeferredSearchTerm] = useState('');
-
-  const [isPending, startTransition] = useTransition();
-
-  const filteredContacts = useMemo(() => contacts.filter((contact) => (
-    contact.name.toLowerCase().includes(DeferredSearchTerm.toLowerCase())
-  )), [contacts, DeferredSearchTerm]);
-
-  const loadContacts = useCallback(async () => {
-    try {
-      setIsLoading(true);
-
-      const contactsList = await ContactsService.listContacts(orderBy);
-
-      setHasError(false);
-      setContacts(contactsList);
-    } catch (error) {
-      setHasError(true);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [orderBy]);
-
-  useEffect(() => {
-    loadContacts();
-  }, [loadContacts]);
-
-  function handleChangeSearchTerm(event) {
-    const { value } = event.target;
-    setSearchTerm(value);
-
-    startTransition(() => {
-      setDeferredSearchTerm(value);
-    });
-  }
-
-  function handleTryAgain() {
-    loadContacts();
-  }
-
-  const handleDeleteContact = useCallback((contact) => {
-    setIsDeleteModalVisible(true);
-    setContactBeingDeleted(contact);
-  }, []);
-
-  function handleCloseDeleteModal() {
-    setIsDeleteModalVisible(false);
-    setContactBeingDeleted(null);
-  }
-
-  const handleOrderBy = useCallback(() => {
-    setOrderBy((prev) => (prev === 'asc' ? 'desc' : 'asc'));
-  }, []);
-
-  async function handleConfirmDeleteContact() {
-    try {
-      setIsLoadingDelete(true);
-
-      await ContactsService.deleteContact(contactBeingDeleted.id);
-
-      setContacts((prev) => prev.filter((contact) => contact.id !== contactBeingDeleted.id));
-
-      handleCloseDeleteModal();
-
-      toast({
-        type: 'success',
-        text: 'Contato excluído com sucesso!',
-      });
-    } catch {
-      toast({
-        type: 'error',
-        text: 'Ocorreu um erro ao excluir o contato!',
-      });
-    } finally {
-      setIsLoadingDelete(false);
-    }
-  }
+  const {
+    isLoading,
+    contactBeingDeleted,
+    handleCloseDeleteModal,
+    handleConfirmDeleteContact,
+    isDeleteModalVisible,
+    isLoadingDelete,
+    contacts,
+    searchTerm,
+    handleChangeSearchTerm,
+    hasError,
+    filteredContacts,
+    isPending,
+    handleTryAgain,
+    orderBy,
+    handleDeleteContact,
+    handleOrderBy,
+  } = useHome();
 
   return (
     <Container>
